fix: re-structure
