+++ packages/pnpmc-show/src/analyze.test.ts
@@ -9,16 +9,14 @@ import {
 } from './analyze.js'
 
 test('collectCatalogDependencies', async () => {
-  const manifest = await readWorkspaceManifest(
-    path.resolve(__dirname, '../../test/fixtures/default')
-  )
+  const manifest = await readWorkspaceManifest(path.resolve(__dirname, '../test/fixtures/default'))
   const deps = collectCatalogDependencies(manifest!)
   expect([...deps]).toEqual(['typescript'])
 })
 
 test('collectProjectDependencies', async () => {
   const projects = await findWorkspacePackages(
-    path.resolve(__dirname, '../../test/fixtures/no-catalog')
+    path.resolve(__dirname, '../test/fixtures/no-catalog')
   )
   const deps = collectProjectDependencies(projects)
   const [root, package1, package2] = projects
@@ -29,7 +27,7 @@ test('collectProjectDependencies', async () => {
 
 describe('analyzeDependencies', () => {
   test('no catalog', async () => {
-    const workspaceDir = path.resolve(__dirname, '../../test/fixtures/no-catalog')
+    const workspaceDir = path.resolve(__dirname, '../test/fixtures/no-catalog')
     const manifest = await readWorkspaceManifest(workspaceDir)
     const deps = await analyzeDependencies(workspaceDir, manifest!)
 
@@ -48,7 +46,7 @@ describe('analyzeDependencies', () => {
   })
 
   test('defined catalog', async () => {
-    const workspaceDir = path.resolve(__dirname, '../../test/fixtures/defined-catalog')
+    const workspaceDir = path.resolve(__dirname, '../test/fixtures/defined-catalog')
     const manifest = await readWorkspaceManifest(workspaceDir)
     const deps = await analyzeDependencies(workspaceDir, manifest!)
 
@@ -56,7 +54,7 @@ describe('analyzeDependencies', () => {
   })
 
   test('defined with named catalog', async () => {
-    const workspaceDir = path.resolve(__dirname, '../../test/fixtures/named')
+    const workspaceDir = path.resolve(__dirname, '../test/fixtures/named')
     const manifest = await readWorkspaceManifest(workspaceDir)
     const deps = await analyzeDependencies(workspaceDir, manifest!)
 
@@ -64,7 +62,7 @@ describe('analyzeDependencies', () => {
   })
 
   test('dep still is not yet in the catalog', async () => {
-    const workspaceDir = path.resolve(__dirname, '../../test/fixtures/still-not-defined-catalog')
+    const workspaceDir = path.resolve(__dirname, '../test/fixtures/still-not-defined-catalog')
     const manifest = await readWorkspaceManifest(workspaceDir)
     const deps = await analyzeDependencies(workspaceDir, manifest!)
 
@@ -83,7 +81,7 @@ describe('analyzeDependencies', () => {
   })
 
   test('dep in root', async () => {
-    const workspaceDir = path.resolve(__dirname, '../../test/fixtures/root')
+    const workspaceDir = path.resolve(__dirname, '../test/fixtures/root')
     const manifest = await readWorkspaceManifest(workspaceDir)
     const deps = await analyzeDependencies(workspaceDir, manifest!)
 
@@ -107,7 +105,7 @@ describe('analyzeDependencies', () => {
   })
 
   test('workspace', async () => {
-    const workspaceDir = path.resolve(__dirname, '../../test/fixtures/workspace')
+    const workspaceDir = path.resolve(__dirname, '../test/fixtures/workspace')
     const manifest = await readWorkspaceManifest(workspaceDir)
     const deps = await analyzeDependencies(workspaceDir, manifest!)
 
@@ -115,7 +113,7 @@ describe('analyzeDependencies', () => {
   })
 
   test('devDependencies', async () => {
-    const workspaceDir = path.resolve(__dirname, '../../test/fixtures/dev-deps')
+    const workspaceDir = path.resolve(__dirname, '../test/fixtures/dev-deps')
     const manifest = await readWorkspaceManifest(workspaceDir)
     const deps = await analyzeDependencies(workspaceDir, manifest!)
