+++ packages/pnpmc-show/src/analyze.ts
