--- packages/pnpmc-show/src/command/index.ts
@@ -1,9 +1,0 @@
-/**
- * @author kazuya kawaguchi (a.k.a @kazupon)
- * @license MIT
- */
-
-import meta from './meta.js'
-import run from './runner.js'
-
-export default { ...meta, run }
