+++ packages/pnpmc-show/tsdown.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'tsdown'
 
 export default defineConfig({
   outDir: 'lib',
-  entry: ['src/index.ts', 'src/command/index.ts', 'src/command/meta.ts', 'src/command/runner.ts'],
+  entry: ['src/cli.ts', 'src/index.ts', 'src/meta.ts', 'src/runner.ts'],
   dts: true,
   clean: true,
   external: ['../package.json']
